Refactor (#4)
* rename folders and add index

* add redirect when logged in

* create page wrapper to remove container-box combo
