+++ src/pages/ChoreCreation/ChoreCreation.jsx
@@ -1,16 +1,10 @@
 import React, { useState } from "react";
-import {
-  Box,
-  Typography,
-  TextField,
-  Button,
-  Stack,
-  Avatar,
-} from "@mui/material";
+import { Typography, TextField, Button, Stack, Avatar } from "@mui/material";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { useNavigate } from "react-router-dom";
+import PageWrapper from "../../components/PageWrapper";
 
 function ChoreCreation() {
   const navigate = useNavigate();
@@ -36,14 +30,7 @@ function ChoreCreation() {
 
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          p: 3,
-        }}
-      >
+      <PageWrapper>
         <Avatar
           src="/EarnNLearn.jpg"
           alt="Logo"
@@ -88,7 +75,7 @@ function ChoreCreation() {
             Cancel
           </Button>
         </Stack>
-      </Box>
+      </PageWrapper>
     </LocalizationProvider>
   );
 }
