+++ src/pages/ProfileSelect/ProfileSelect.jsx
@@ -4,6 +4,7 @@ import { Link } from "react-router-dom";
 import AddIcon from "@mui/icons-material/Add";
 import { callApi } from "../../utils/api.util";
 import { METHOD } from "../../constants/enums";
+import PageWrapper from "../../components/PageWrapper";
 
 function ProfileSelect() {
   // Dummy data for kids' profiles with unique colors
@@ -58,67 +59,58 @@ function ProfileSelect() {
   }
 
   return (
-    <Container component="main" maxWidth="sm">
-      <Box
-        sx={{
-          marginTop: 8,
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-        }}
-      >
-        <Avatar
-          src="/EarnNLearn.jpg"
-          alt="Logo"
-          sx={{ width: 100, height: 100, marginBottom: 2 }}
-        />
-        <Typography component="h1" variant="h4" sx={{ marginBottom: 3 }}>
-          Profiles
-        </Typography>
-        {children.map((child, index) => (
-          <Link
-            key={child.id}
-            to={`/profile-balance/${child.id}`}
-            style={{ textDecoration: "none", width: "100%" }}
-          >
-            <Box
-              sx={{
-                display: "flex",
-                alignItems: "center",
-                padding: 2,
-                marginBottom: 1,
-                border: "1px solid #ccc",
-                borderRadius: "5px",
-                height: 60,
-                "&:hover": {
-                  backgroundColor: "#f0f0f0",
-                  cursor: "pointer",
-                },
-              }}
-            >
-              <Avatar
-                sx={{
-                  marginRight: 2,
-                  backgroundColor: colors[index % colors.length],
-                }}
-              />
-              <Typography variant="h6">
-                {child.firstName} {child.lastName}
-              </Typography>
-            </Box>
-          </Link>
-        ))}
-        <Button
-          variant="contained"
-          sx={{ marginTop: 2 }}
-          startIcon={<AddIcon />}
-          component={Link}
-          to="/add-profile"
+    <PageWrapper>
+      <Avatar
+        src="/EarnNLearn.jpg"
+        alt="Logo"
+        sx={{ width: 100, height: 100, marginBottom: 2 }}
+      />
+      <Typography component="h1" variant="h4" sx={{ marginBottom: 3 }}>
+        Profiles
+      </Typography>
+      {children.map((child, index) => (
+        <Link
+          key={child.id}
+          to={`/profile-balance/${child.id}`}
+          style={{ textDecoration: "none", width: "100%" }}
         >
-          Add Profile
-        </Button>
-      </Box>
-    </Container>
+          <Box
+            sx={{
+              display: "flex",
+              alignItems: "center",
+              padding: 2,
+              marginBottom: 1,
+              border: "1px solid #ccc",
+              borderRadius: "5px",
+              height: 60,
+              "&:hover": {
+                backgroundColor: "#f0f0f0",
+                cursor: "pointer",
+              },
+            }}
+          >
+            <Avatar
+              sx={{
+                marginRight: 2,
+                backgroundColor: colors[index % colors.length],
+              }}
+            />
+            <Typography variant="h6">
+              {child.firstName} {child.lastName}
+            </Typography>
+          </Box>
+        </Link>
+      ))}
+      <Button
+        variant="contained"
+        sx={{ marginTop: 2 }}
+        startIcon={<AddIcon />}
+        component={Link}
+        to="/add-profile"
+      >
+        Add Profile
+      </Button>
+    </PageWrapper>
   );
 }
