+++ src/pages/login/Login.jsx
@@ -1,7 +1,6 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import {
-  Container,
   TextField,
   Button,
   Typography,
@@ -12,11 +11,21 @@ import {
 import "./LoginStyles.css";
 import { callApi } from "../../utils/api.util";
 import { METHOD } from "../../constants/enums";
+import { getCookie } from "../../utils/auth.util";
+import PageWrapper from "../../components/PageWrapper";
 
 function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const navigate = useNavigate();
+
+  useEffect(() => {
+    const isAuth = getCookie("jwt");
+
+    if (isAuth) {
+      navigate("/profiles");
+    }
+  }, []);
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -43,82 +52,73 @@ function Login() {
   };
 
   return (
-    <Container component="main" maxWidth="xs">
-      <Box
-        sx={{
-          marginTop: 8,
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-        }}
+    <PageWrapper>
+      <Avatar
+        src="/EarnNLearn.jpg"
+        alt="Logo"
+        sx={{ width: 100, height: 100, marginBottom: 2 }} // Adjust size as needed
+      />
+      <Typography
+        component="h1"
+        variant="h5"
+        sx={{ marginBottom: 2, color: "hotpink" }}
       >
-        <Avatar
-          src="/EarnNLearn.jpg"
-          alt="Logo"
-          sx={{ width: 100, height: 100, marginBottom: 2 }} // Adjust size as needed
+        EarnNLearn
+      </Typography>
+      <Box component="form" noValidate className="form">
+        <InputLabel htmlFor="email">Email</InputLabel>
+        <TextField
+          margin="normal"
+          required
+          fullWidth
+          id="email"
+          name="email"
+          autoComplete="email"
+          autoFocus
+          className="textField"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
         />
-        <Typography
-          component="h1"
-          variant="h5"
-          sx={{ marginBottom: 2, color: "hotpink" }}
-        >
-          EarnNLearn
+        <InputLabel htmlFor="password">Password</InputLabel>
+        <TextField
+          margin="normal"
+          required
+          fullWidth
+          name="password"
+          type="password"
+          id="password"
+          autoComplete="current-password"
+          className="textField"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+        <Typography className="accountPrompt">
+          Don't have an account?{" "}
+          <Link to="/register" className="linkText">
+            Register.
+          </Link>
         </Typography>
-        <Box component="form" noValidate className="form">
-          <InputLabel htmlFor="email">Email</InputLabel>
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            id="email"
-            name="email"
-            autoComplete="email"
-            autoFocus
-            className="textField"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-          <InputLabel htmlFor="password">Password</InputLabel>
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            name="password"
-            type="password"
-            id="password"
-            autoComplete="current-password"
-            className="textField"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <Typography className="accountPrompt">
-            Don't have an account?{" "}
-            <Link to="/register" className="linkText">
-              Register.
-            </Link>
-          </Typography>
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            className="signInButton"
-            sx={{ mt: 3, mb: 2, backgroundColor: "#0D99FF" }}
-            onClick={(e) => handleSubmit(e)}
-          >
-            Sign In
-          </Button>
-          {/*route to the Profiles page for testing */}
-          <Button
-            component={Link}
-            to="/profiles"
-            variant="contained"
-            sx={{ mt: 3, mb: 2 }}
-          >
-            Go to Profiles
-          </Button>
-        </Box>
+        <Button
+          type="submit"
+          fullWidth
+          variant="contained"
+          className="signInButton"
+          sx={{ mt: 3, mb: 2, backgroundColor: "#0D99FF" }}
+          onClick={(e) => handleSubmit(e)}
+        >
+          Sign In
+        </Button>
+        {/*route to the Profiles page for testing */}
+        <Button
+          component={Link}
+          to="/profiles"
+          variant="contained"
+          sx={{ mt: 3, mb: 2 }}
+        >
+          Go to Profiles
+        </Button>
       </Box>
-    </Container>
+    </PageWrapper>
   );
 }
